pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game-bot-mumu"
version = "0.1.0"
description = "Python game auto-farm bot for Android emulators (MuMu first)."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "PySide6>=6.7",
  "opencv-python>=4.10",
  "numpy>=1.26",
  "PyYAML>=6.0",
  "paddleocr>=2.8",
]

[dependency-groups]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "-p no:cacheprovider"
markers = ["integration: tests requiring real adb/emulator"]
